dspy_service/requirements.txt
flask
openai
litellm>=1.35.3
dspy>=3.0.4b2
faiss-cpu
numpy
python-dotenv